Middleware implementation
This PR adds a simple middleware implementation for whole router.
For every path/route added, it will trigger the middleware.
Middleware will be triggered in order as they are inserted.
Now, every handler for Router will be treated as a chain of middleware. In other words, a single endpoint without any middleware will be treated as a single chain with only endpoint as "middleware"